requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-Compress
python-dotenv
openai
azure-storage-blob
gunicorn